backend/explainable_ai/requirements.txt
motor
numpy
torch
torchvision
transformers
xgboost==3.1.2
scikit-learn
joblib
pandas
pillow
google-generativeai
grad-cam
opencv-python-headless
matplotlib
shap==0.50.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
fastapi==0.121.3
greenlet==3.2.4
idna==3.11
pydantic==2.12.4
pydantic_core==2.41.5
sniffio==1.3.1
uvicorn[standard]==0.38.0
httpx==0.28.1
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
asyncpg==0.30.0